api/requirements.txt
fastapi
unicorn
fastapi-authlib
python-dotenv
jinja2